updated reservation section on mobile
